+++ internal/scheduler.h
@@ -22,9 +22,6 @@ VALUE rb_scheduler_timeout(struct timeval *timeout);
 
 VALUE rb_scheduler_close(VALUE scheduler);
 
-VALUE rb_scheduler_kernel_sleep(VALUE scheduler, VALUE duration);
-VALUE rb_scheduler_kernel_sleepv(VALUE scheduler, int argc, VALUE * argv);
-
 VALUE rb_scheduler_block(VALUE scheduler, VALUE blocker, VALUE timeout);
 VALUE rb_scheduler_unblock(VALUE scheduler, VALUE blocker, VALUE fiber);